requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
flask-restplus==0.13.0
Flask-WTF==0.14.2
gunicorn==19.9.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-speech==1.2.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.0
idna==2.8
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.6
flask-cors
